+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/favorite/FavoriteViewModel.kt
@@ -1,27 +1,25 @@
-package com.sw.sw_api_kotlin_project.fragment.planet
+package com.sw.sw_api_kotlin_project.screen.favorite
 
 import androidx.lifecycle.liveData
 import com.sw.sw_api_kotlin_project.base.BaseViewModel
-import com.sw.sw_api_kotlin_project.repository.PlanetRepository
-import com.sw.sw_api_kotlin_project.utils.PageType
+import com.sw.sw_api_kotlin_project.repository.FavoriteRepository
 import com.sw.sw_api_kotlin_project.utils.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
 
 @HiltViewModel
-class PlanetViewModel @Inject constructor(
-    private val planetRepository: PlanetRepository
+class FavoriteViewModel @Inject constructor(
+    private val favoriteRepository: FavoriteRepository
 ) : BaseViewModel() {
 
-    fun getPlanets(pageType: PageType) = liveData(Dispatchers.IO) {
-        pageParameterFormat(pageType)
-        emit(Resource.loading(null))
+    fun getFavoriteList() = liveData(Dispatchers.IO) {
         try {
-            val response = planetRepository.getPlanets(page)
+            val response = favoriteRepository.getAll()
             emit(Resource.success(response))
         } catch (e: Exception) {
             emit(Resource.error(data = null, message = e.message ?: "error"))
         }
     }
+
 }
