+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.fragment.home
+package com.sw.sw_api_kotlin_project.screen.home
 
 import com.sw.sw_api_kotlin_project.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
