+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/others/OthersFragment.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.fragment.others
+package com.sw.sw_api_kotlin_project.screen.others
 
 import android.os.Bundle
 import android.view.LayoutInflater
