+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/people/PeopleListViewModel.kt
@@ -1,21 +1,24 @@
-package com.sw.sw_api_kotlin_project.fragment.favorite
+package com.sw.sw_api_kotlin_project.screen.people
 
 import androidx.lifecycle.liveData
 import com.sw.sw_api_kotlin_project.base.BaseViewModel
-import com.sw.sw_api_kotlin_project.repository.FavoriteRepository
+import com.sw.sw_api_kotlin_project.repository.PeopleRepository
+import com.sw.sw_api_kotlin_project.utils.PageType
 import com.sw.sw_api_kotlin_project.utils.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
 
 @HiltViewModel
-class FavoriteViewModel @Inject constructor(
-    private val favoriteRepository: FavoriteRepository
+class PeopleListViewModel @Inject constructor(
+    private val peopleRepository: PeopleRepository
 ) : BaseViewModel() {
 
-    fun getFavoriteList() = liveData(Dispatchers.IO) {
+    fun getPeople(pageType: PageType) = liveData(Dispatchers.IO) {
+        pageParameterFormat(pageType)
+        emit(Resource.loading(null))
         try {
-            val response = favoriteRepository.getAll()
+            val response = peopleRepository.getPeople(page)
             emit(Resource.success(response))
         } catch (e: Exception) {
             emit(Resource.error(data = null, message = e.message ?: "error"))
