+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/planet/PlanetViewModel.kt
@@ -1,8 +1,8 @@
-package com.sw.sw_api_kotlin_project.fragment.people
+package com.sw.sw_api_kotlin_project.screen.planet
 
 import androidx.lifecycle.liveData
 import com.sw.sw_api_kotlin_project.base.BaseViewModel
-import com.sw.sw_api_kotlin_project.repository.PeopleRepository
+import com.sw.sw_api_kotlin_project.repository.PlanetRepository
 import com.sw.sw_api_kotlin_project.utils.PageType
 import com.sw.sw_api_kotlin_project.utils.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -10,19 +10,18 @@ import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
 
 @HiltViewModel
-class PeopleListViewModel @Inject constructor(
-    private val peopleRepository: PeopleRepository
+class PlanetViewModel @Inject constructor(
+    private val planetRepository: PlanetRepository
 ) : BaseViewModel() {
 
-    fun getPeople(pageType: PageType) = liveData(Dispatchers.IO) {
+    fun getPlanets(pageType: PageType) = liveData(Dispatchers.IO) {
         pageParameterFormat(pageType)
         emit(Resource.loading(null))
         try {
-            val response = peopleRepository.getPeople(page)
+            val response = planetRepository.getPlanets(page)
             emit(Resource.success(response))
         } catch (e: Exception) {
             emit(Resource.error(data = null, message = e.message ?: "error"))
         }
     }
-
 }
