+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/search/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.fragment.search
+package com.sw.sw_api_kotlin_project.screen.search
 
 import android.os.Bundle
 import android.os.Parcelable
