+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_film_list.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="fragment_films"
-            type="com.sw.sw_api_kotlin_project.fragment.film.FilmListViewModel" />
+            type="com.sw.sw_api_kotlin_project.screen.film.FilmListViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
