+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_planet_details.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.sw.sw_api_kotlin_project.fragment.planet_details.PlanetDetailsViewModel" />
+            type="com.sw.sw_api_kotlin_project.screen.planet_details.PlanetDetailsViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
