+++ SW_API_Kotlin_Project/app/src/main/res/navigation/nav_main_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/nav_home_list"
-        android:name="com.sw.sw_api_kotlin_project.fragment.home.HomeFragment"
+        android:name="com.sw.sw_api_kotlin_project.screen.home.HomeFragment"
         android:label="HomeFragment"
         tools:layout="@layout/fragment_home">
         <action
@@ -41,7 +41,7 @@
 
     <fragment
         android:id="@+id/nav_search"
-        android:name="com.sw.sw_api_kotlin_project.fragment.search.SearchFragment"
+        android:name="com.sw.sw_api_kotlin_project.screen.search.SearchFragment"
         android:label="SearchFragment"
         tools:layout="@layout/fragment_search">
         <action
@@ -57,7 +57,7 @@
 
     <fragment
         android:id="@+id/nav_favorite_list"
-        android:name="com.sw.sw_api_kotlin_project.fragment.favorite.FavoriteFragment"
+        android:name="com.sw.sw_api_kotlin_project.screen.favorite.FavoriteFragment"
         android:label="FavoriteFragment"
         tools:layout="@layout/fragment_favorite">
         <action
@@ -73,7 +73,7 @@
 
     <fragment
         android:id="@+id/nav_other"
-        android:name="com.sw.sw_api_kotlin_project.fragment.others.OthersFragment"
+        android:name="com.sw.sw_api_kotlin_project.screen.others.OthersFragment"
         android:label="OthersFragment"
         tools:layout="@layout/fragment_others">
         <action
